Rename dictionary to constants
It makes it consistent with standard naming conventions
