+++ eva/configuration/configuration_manager.py
@@ -16,7 +16,7 @@ import yaml
 
 from eva.configuration.bootstrap_environment import bootstrap_environment
 from eva.configuration.config_utils import read_value_config, update_value_config
-from eva.configuration.dictionary import (
+from eva.configuration.constants import (
     EVA_CONFIG_FILE,
     EVA_DEFAULT_DIR,
     EVA_INSTALLATION_DIR,
