+++ test/util.py
@@ -23,7 +23,7 @@ import pandas as pd
 from eva.binder.statement_binder import StatementBinder
 from eva.binder.statement_binder_context import StatementBinderContext
 from eva.configuration.configuration_manager import ConfigurationManager
-from eva.configuration.dictionary import EVA_INSTALLATION_DIR
+from eva.configuration.constants import EVA_INSTALLATION_DIR
 from eva.models.catalog.frame_info import FrameInfo
 from eva.models.catalog.properties import ColorSpace
 from eva.models.storage.batch import Batch
